LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scope-shadow-all-expected.txt


FAIL anchor-scope:all is tree-scoped assert_equals: expected "10px" but got "1px"
